requirements.txt
# LLM APIs
anthropic==0.84.0
aider-chat==0.86.2
backoff==2.2.1
openai==2.20.0
google-generativeai==0.8.6

# Literature / HTTP
requests==2.32.5
pyalex==0.21

# PDF / Writeup
pypdf==6.7.5
pymupdf==1.27.1
pymupdf4llm==0.3.4

# Visualization
matplotlib==3.10.8

# Remote execution
paramiko>=3.4
pyyaml>=6.0

# Core ML
torch==2.10.0
numpy==2.4.2
transformers==5.2.0
datasets==4.6.1
tiktoken==0.12.0
wandb==0.25.0
tqdm==4.67.3
